+++ src/test/java/de/alpharogroup/file/filter/TxtFileFilterTest.java
@@ -36,7 +36,7 @@ import java.nio.charset.Charset;
 import org.apache.commons.io.FileUtils;
 import org.testng.annotations.Test;
 
-import de.alpharogroup.evaluate.object.ToStringEvaluator;
+import de.alpharogroup.evaluate.object.evaluators.ToStringEvaluator;
 import de.alpharogroup.lang.ClassExtensions;
 
 /**