+++ src/test/java/de/alpharogroup/lang/util/ManifestVersionTest.java
@@ -32,7 +32,7 @@ import java.util.jar.Attributes;
 import org.meanbean.test.BeanTester;
 import org.testng.annotations.Test;
 
-import de.alpharogroup.evaluate.object.SilentEqualsHashCodeAndToStringEvaluator;
+import de.alpharogroup.evaluate.object.evaluators.SilentEqualsHashCodeAndToStringEvaluator;
 
 /**
  * The unit test class for the class {@link ManifestVersion}.